requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
anyascii==0.3.2
    # via wagtail
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
beautifulsoup4==4.11.2
    # via wagtail
black==23.12.1
    # via -r requirements/dev.in
brotli==1.1.0
    # via fonttools
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via weasyprint
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
crispy-bootstrap5==2023.10
    # via -r requirements/base.in
cssselect2==0.7.0
    # via weasyprint
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   odfpy
    #   willow
diff-match-patch==20230430
    # via django-import-export
django==4.2.9
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap5
    #   django-ace
    #   django-braces
    #   django-comments-xtd
    #   django-contrib-comments
    #   django-crispy-forms
    #   django-filter
    #   django-import-export
    #   django-model-utils
    #   django-modelcluster
    #   django-permissionedforms
    #   django-star-ratings
    #   django-stubs
    #   django-stubs-ext
    #   django-tables2
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-ace==1.32.0
    # via -r requirements/base.in
django-braces==1.15.0
    # via django-star-ratings
django-colorfield==0.11.0
    # via -r requirements/base.in
django-comments-xtd==2.9.13
    # via -r requirements/base.in
django-contrib-comments==2.2.0
    # via django-comments-xtd
django-crispy-forms==2.1
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap5
django-etc==1.4.0
    # via django-hitcount
django-filter==23.5
    # via wagtail
django-hitcount==1.3.5
    # via -r requirements/base.in
django-import-export==3.3.5
    # via -r requirements/base.in
django-model-utils==4.3.1
    # via django-star-ratings
django-modelcluster==6.2
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-star-ratings==0.9.2
    # via -r requirements/base.in
django-stubs==4.2.7
    # via -r requirements/dev.in
django-stubs-ext==4.2.7
    # via django-stubs
django-tables2==2.7.0
    # via -r requirements/base.in
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
django-treenode==0.21.0
    # via -r requirements/base.in
djangorestframework==3.14.0
    # via
    #   django-comments-xtd
    #   wagtail
docutils==0.20.1
    # via django-comments-xtd
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
executing==2.0.1
    # via stack-data
filetype==1.2.0
    # via willow
fonttools[woff]==4.47.0
    # via
    #   fonttools
    #   weasyprint
html5lib==1.1
    # via
    #   wagtail
    #   weasyprint
idna==3.6
    # via requests
ipython==8.19.0
    # via -r requirements/dev.in
isort==5.13.2
    # via -r requirements/dev.in
jedi==0.19.1
    # via ipython
l18n==2021.3
    # via wagtail
lxml==5.0.0
    # via -r requirements/base.in
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.6
    # via ipython
mypy-extensions==1.0.0
    # via black
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via
    #   -r requirements/base.in
    #   tablib
    #   wagtail
packaging==23.2
    # via
    #   black
    #   build
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   -r requirements/base.in
    #   django-colorfield
    #   pillow-heif
    #   wagtail
    #   weasyprint
pillow-heif==0.14.0
    # via willow
pip-tools==7.3.0
    # via -r requirements/dev.in
platformdirs==4.1.0
    # via black
prompt-toolkit==3.0.43
    # via ipython
psycopg[c]==3.1.16
    # via -r requirements/base.in
psycopg-c==3.1.16
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pycryptodome==3.19.1
    # via -r requirements/base.in
pydyf==0.8.0
    # via weasyprint
pygments==2.17.2
    # via ipython
pyphen==0.14.0
    # via weasyprint
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via -r requirements/base.in
pytz==2023.3.post1
    # via
    #   django-comments-xtd
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via tablib
requests==2.31.0
    # via wagtail
six==1.16.0
    # via
    #   asttokens
    #   django-comments-xtd
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
swapper==1.3.0
    # via django-star-ratings
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   django-import-export
    #   tablib
telepath==0.3.1
    # via wagtail
tinycss2==1.2.1
    # via
    #   cssselect2
    #   weasyprint
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
types-pytz==2023.3.1.1
    # via django-stubs
types-pyyaml==6.0.12.12
    # via django-stubs
typing-extensions==4.9.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   psycopg
tzdata==2023.4
    # via -r requirements/base.in
urllib3==2.1.0
    # via requests
wagtail==5.2.2
    # via -r requirements/base.in
wcwidth==0.2.12
    # via prompt-toolkit
weasyprint==60.2
    # via -r requirements/base.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.42.0
    # via pip-tools
willow[heif]==1.6.3
    # via
    #   wagtail
    #   willow
xlrd==2.0.1
    # via
    #   -r requirements/base.in
    #   tablib
xlwt==1.3.0
    # via tablib
zopfli==0.2.3
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
